[WAGON-338] Exception from cyberneko from jenkins repo listing
the html listing parser was coded to use Cyberneko. This orphaned ball of complexity crashes on some things
it does not understand, including some output from the Jenkins CI system when it is serving as a repository.

A maintained, superior alternative is Jsoup. This change switches wagon-http-shared4 to use it.
It passes all the tests.


git-svn-id: https://svn.apache.org/repos/asf/maven/wagon/trunk@1143141 13f79535-47bb-0310-9956-ffa450edef68