[WAGON-529] Clean up inconsistent status code and reason phrase handling
* Write a cleaned up message to TransferFailedException
* Drop reason phrase from other exceptions where the message already
contains the phrase
* Renamed the HttpWagonReasonPhraseTest to HttpWagonErrorTest because it
was ill-designed: the test assumed that the message set by
HttpServletResponse#sendError() is returned in the status line to the
client by the server, but this is not true. The Servlet Spec says that
this message (not reason) has to be written to an HTML page. In this
case, Jetty simply wrote that to the status line, but we shall not rely
on an implementation detail of just one server. The spec does not offer
an option to actively set the reason phrase.

The reason phrase is optional after all, don't use it. Use the status
code.