[WAGON-568] Fail to deploy on Sonatype OSS since Maven 3.5.4
Previously we assumed with manual relocation that a request can always be
replayed (repeated). This caused stream-based resources with a provided
content length the peer to block for advertised bytes until a
REQUEST_TIMEOUT (408) is issued.

Now we properly detect non-repeatable streams and abort any redirect in such
a case. A new test will verify a TransferFailedException for this case.

See also WAGON-570, WAGON-581 (HTTPCLIENT-2052).